chore: move bus and topic handler to logic packages
