+++ pkg/costs/deployer/config_refresher.go
@@ -11,6 +11,7 @@ import (
 	corev1 "k8s.io/client-go/kubernetes/typed/core/v1"
 	"k8s.io/client-go/rest"
 
+	"github.com/seal-io/seal/pkg/bus/connector"
 	"github.com/seal-io/seal/pkg/dao/model"
 	"github.com/seal-io/seal/pkg/dao/types"
 	"github.com/seal-io/seal/pkg/dao/types/status"
@@ -96,4 +97,13 @@ func refreshCustomPricing(conn *model.Connector, restCfg *rest.Config) error {
 		return fmt.Errorf("error response from %s, expected %d but get code: %d", url, http.StatusOK, resp.StatusCode)
 	}
 	return nil
+}
+
+func SyncCostCustomPricing(ctx context.Context, message connector.BusMessage) error {
+	conn := message.Refer
+	if !conn.EnableFinOps {
+		return nil
+	}
+
+	return UpdateCustomPricing(ctx, conn)
 }
